requirements.txt
bentoml>=1.2.4
torch==2.1.2
transformers==4.38.2
triton==2.1.0; sys_platform == "linux"
TTS==0.22.0
accelerate==0.26.1
diffusers==0.25.1
pillow==10.2.0
torch==2.1.2
xformers==0.0.23.post1
moviepy
ffmpeg-python
vllm==0.3.3
asyncio